+++ Assets/Prefabs/Enemy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 38b01a6a2a3a4ea44b67b676e427ee6e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
