+++ Assets/Prefabs/Enemy/Shooter.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: c8c114691923bc14d98548a51ea94e69
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
