+++ Assets/Prefabs/Spawnables.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6f3642818d0885b4c87be89e62d21620
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
