+++ Assets/Scripts/CameraScripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5b3eb3520c49d854fbe481f7e9374ee2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
