+++ Assets/Scripts/EnvironmentScripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2c544dd7214385c4f9bbcf4701afbb3a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
