+++ Assets/Scripts/GlobalScripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e9dd97478c96038468f8e959a5d59295
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
