+++ Assets/Scripts/GlobalScripts/EnemySpawnerScript.cs
@@ -1,0 +1,20 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class EnemySpawnerScript : MonoBehaviour
+{
+    public GameObject enemy;
+    // Start is called before the first frame update
+    void Start()
+    {
+        
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        if (Input.GetKeyDown(KeyCode.Space))
+        Instantiate(enemy, transform.position, transform.rotation);
+    }
+}
