+++ Assets/Scripts/GlobalScripts/EnemySpawnerScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f12579e68a1abb144aa740c83c033165
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
