+++ Assets/Scripts/PlayerScripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b653ce5133e929d48bee059ba6a53256
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
