+++ Assets/Scripts/ProjectileScripts/BulletScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b9c6f9311361d7940b7351146a82295b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
