+++ Assets/Scripts/UIScripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 44ac6d1b021aea54cabd636dc38aa273
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
