+++ Assets/Sprites/Player.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2b5cda1283856b34c9f3908098fc0c11
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
