+++ Assets/Sprites/UI.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b0690bb14456f9541b76c8d089714c00
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
